Правки по предыдущему разделу и редактирование render функции
