+++ src/const.js
@@ -1,12 +1,5 @@
 export const POINT_COUNT = 18;
 
-export const RenderPosition = {
-  BEFOREBEGIN: 'beforebegin',
-  AFTERBEGIN: 'afterbegin',
-  BEFOREEND: 'beforeend',
-  AFTEREND: 'afterend',
-};
-
 export const PointStateFilter = {
   EVERYTHING: 'everything',
   FUTURE: 'future',
@@ -31,45 +24,4 @@ export const pointTypes = [
   'check-in',
   'sightseeing',
   'restaurant',
-];
-
-export const Destination = {
-  Amsterdam: 'Amsterdam',
-  Cheboksary: 'Cheboksary',
-  Qatar: 'Qatar',
-  Dubai: 'Dubai',
-  Moscow: 'Moscow',
-  Ottawa: 'Ottawa',
-};
-
-export const destinationTexts = [
-  'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget.',
-  'Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra.',
-  'Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.',
-  'Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. Sed sed nisi sed augue convallis suscipit in sed felis.',
-  'Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus. In rutrum ac purus sit amet tempus.'
-];
-
-export const destinationPhotos = [
-  'img/photos/1.jpg',
-  'img/photos/2.jpg',
-  'img/photos/3.jpg',
-  'img/photos/4.jpg',
-  'img/photos/5.jpg',
-];
-
-export const titleOffers = [
-  'Upgrade to a business class',
-  'Choose the radio station',
-  'Add luggage',
-  'Switch to comfort',
-  'Add breakfast',
-];
-
-export const priceOffers = [
-  '69',
-  '89',
-  '68',
-  '98',
-  '86',
 ];
